requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.0
numpy==1.26.3
sqlalchemy==2.0.25
selenium==4.17.2
gspread==6.0.2
gspread-dataframe==3.3.1
oauth2client==4.1.3
dateparser==1.2.0
lxml==5.1.0
openpyxl==3.1.2